+++ packages/lexical-code/src/HighlighterHelper.ts
@@ -1,0 +1,73 @@
+/**
+ * Copyright (c) Meta Platforms, Inc. and affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ */
+
+// eslint-disable-next-line simple-import-sort/imports
+import {$isLineBreakNode, LexicalNode, LexicalEditor} from 'lexical';
+import {CodeHighlightNode, $isCodeHighlightNode, CodeNode} from '@lexical/code';
+
+export function getFirstCodeHighlightNodeOfLine(
+  anchor: LexicalNode,
+): CodeHighlightNode | null | undefined {
+  let currentNode = null;
+  const previousSiblings = anchor.getPreviousSiblings();
+  previousSiblings.push(anchor);
+  while (previousSiblings.length > 0) {
+    const node = previousSiblings.pop();
+    if ($isCodeHighlightNode(node)) {
+      currentNode = node;
+    }
+    if ($isLineBreakNode(node)) {
+      break;
+    }
+  }
+
+  return currentNode;
+}
+
+export function getLastCodeHighlightNodeOfLine(
+  anchor: LexicalNode,
+): CodeHighlightNode | null | undefined {
+  let currentNode = null;
+  const nextSiblings = anchor.getNextSiblings();
+  nextSiblings.unshift(anchor);
+  while (nextSiblings.length > 0) {
+    const node = nextSiblings.shift();
+    if ($isCodeHighlightNode(node)) {
+      currentNode = node;
+    }
+    if ($isLineBreakNode(node)) {
+      break;
+    }
+  }
+
+  return currentNode;
+}
+
+export function updateCodeGutter(node: CodeNode, editor: LexicalEditor): void {
+  const codeElement = editor.getElementByKey(node.getKey());
+  if (codeElement === null) {
+    return;
+  }
+  const children = node.getChildren();
+  const childrenLength = children.length;
+  // @ts-ignore: internal field
+  if (childrenLength === codeElement.__cachedChildrenLength) {
+    // Avoid updating the attribute if the children length hasn't changed.
+    return;
+  }
+  // @ts-ignore:: internal field
+  codeElement.__cachedChildrenLength = childrenLength;
+  let gutter = '1';
+  let count = 1;
+  for (let i = 0; i < childrenLength; i++) {
+    if ($isLineBreakNode(children[i])) {
+      gutter += '\n' + ++count;
+    }
+  }
+  codeElement.setAttribute('data-gutter', gutter);
+}
